Ran CLIG. Many modifications to accel.h and its functions/structures.
git-svn-id: file://localhost/data1/git/presto_svn/trunk@449 5fac236c-a0d8-0310-9745-ef116914eda1

+++ include/prepfold.h
@@ -62,7 +62,7 @@ int read_floats(FILE *file, float *data, int numpts, int numchan);
 /* point data.                                              */
 /* It returns the number of points read.                    */
 
-void hunt(double *xx, unsigned long n, double x, unsigned long *jlo);
+void hunt(double *xx, int n, double x, int *jlo);
 
 int dgels_(char *trans, int *mm, int *nn, int *nrhs, 
 	   double *aa, int *lda, double *bb, int *ldb,
